+++ NEWS.md
@@ -1,5 +1,8 @@
 # ctrdata 1.25.0.9000
 
+## Enhancement
+- Work around upstream issue (#59, CTIS delivering at most 10000 trials per query)
+
 ## Bug fix
 - Re-added field missing from EUCTR conversion
 - Changed API call after long deprecated by ISRCTN